requirements.txt
google-adk>=0.1.0
litellm>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.0.0